+++ src/main/java/com/vizcainoraw/hotelalura/dto/costumer/CustomerDto.java
@@ -4,10 +4,9 @@ package com.vizcainoraw.hotelalura.dto.costumer;
  * CustomerDto
  */
 public record CustomerDto(
-    Integer id,
+    Long id,
     String email,
     String firstName,
     String lastName,
     String phoneNumber
-) {
-}
+) {}
